healthcare-cloud-system/microservices/user-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic[email]==2.5.0
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
kafka-python==2.0.2
prometheus-client==0.19.0
python-multipart==0.0.6
